Move purge_zones_without_loss_points into aggregate_loss_by_zone.py
